Configures separate production and sample databases
Sets up distinct database configurations for production and sample environments.

Introduces a sample database alongside the production database for development and testing. This includes corresponding environment variables, docker-compose configurations, and database connection logic.

Updates the Makefile to include commands for running and cleaning both production and sample database environments, as well as seeding data for the sample database.
