--- .env.example
@@ -1,4 +1,0 @@
-POSTGRES_USER=cs348
-POSTGRES_PASSWORD=group10
-POSTGRES_DB=housing_db
-DATABASE_URL=postgresql://cs348:group10@localhost:5432/housing_db
