+++ .env.production
@@ -1,0 +1,7 @@
+DB_ENV=production
+
+# Production
+POSTGRES_USER=cs348
+POSTGRES_PASSWORD=group10
+POSTGRES_DB=housing_db
+DATABASE_URL=postgresql://cs348:group10@db:5432/housing_db
