+++ .env.sample
@@ -1,0 +1,13 @@
+DB_ENV=sample
+
+# Production (for Client)
+POSTGRES_USER=cs348
+POSTGRES_PASSWORD=group10
+POSTGRES_DB=housing_db
+DATABASE_URL=postgresql://cs348:group10@db:5432/housing_db
+
+# Sample
+SAMPLE_POSTGRES_USER=cs348-sample
+SAMPLE_POSTGRES_PASSWORD=group10-sample
+SAMPLE_POSTGRES_DB=sample_housing_db
+SAMPLE_DATABASE_URL=postgresql://cs348-sample:group10-sample@db-sample:5432/sample_housing_db
