+++ Makefile
@@ -1,0 +1,31 @@
+# Instead of 'docker-compose up --build', you may use
+#	make run-sample 	to use sample DB
+# 	make run-prod		to use production DB
+#
+# Client (Frontend) will use production DB only in any case
+
+run-sample:
+	DB_ENV=sample docker compose -f docker-compose.yml -f docker-compose.sample.yml --env-file .env.sample up --build
+
+run-prod:
+	DB_ENV=production docker compose -f docker-compose.yml -f docker-compose.production.yml --env-file .env.production up --build
+
+clean-sample:
+	docker compose -f docker-compose.yml -f docker-compose.sample.yml down --volumes --remove-orphans
+
+clean-prod:
+	docker compose -f docker-compose.yml -f docker-compose.production.yml down --volumes --remove-orphans
+
+clean-default:
+	docker compose down --volumes --remove-orphans
+
+clean-all:
+	docker compose -f docker-compose.yml -f docker-compose.sample.yml down --volumes --remove-orphans 2>/dev/null || true
+	docker compose -f docker-compose.yml -f docker-compose.production.yml down --volumes --remove-orphans 2>/dev/null || true
+	docker compose down --volumes --remove-orphans 2>/dev/null || true
+
+run-sample-sql:
+	DB_ENV=sample docker compose -f docker-compose.yml -f docker-compose.sample.yml --env-file .env.sample up -d --build
+	docker compose exec backend python app/seed_data.py
+	docker compose exec -T db-sample psql -U cs348-sample -d sample_housing_db < test-sample.sql > test-sample.out
+	cat test-sample.out
