+++ backend/app/db/__init__.py
@@ -1,0 +1,3 @@
+from .base import Base
+from .db_prod import SessionLocal, engine
+from .db_sample import SessionLocal, engine
