+++ backend/app/db/db_prod.py
@@ -1,14 +1,21 @@
-from sqlalchemy import create_engine, Column, Integer, String
+from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 import os
 
-DATABASE_URL = os.getenv("DATABASE_URL", "postgresql://cs348:group10@localhost:5432/housing_db")
+# production
+DB_ENV = os.getenv("DB_ENV")
 
+def get_prod_db_url():
+    print(f"""
+Run Production Database
+Current Env: {DB_ENV}
+    """)
+    return os.getenv("DATABASE_URL")
+
+DATABASE_URL = get_prod_db_url()
 engine = create_engine(DATABASE_URL)
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-
-Base = declarative_base()
 
 def get_db():
     db = SessionLocal()
@@ -16,3 +23,4 @@ def get_db():
         yield db
     finally:
         db.close()
+
