+++ backend/app/models/property_model.py
@@ -5,7 +5,7 @@ Property(property_id Int, region_id Int, type String, subtype String)
 '''
 
 from sqlalchemy import Column, Integer, String, Float, ForeignKey
-from app.database import Base
+from app.db.base import Base
 
 # Property(property_id, region_id, type, subtype) (Minjae Lee)
