+++ backend/app/models/region_model.py
@@ -1,5 +1,5 @@
 from sqlalchemy import Column, Integer, String
-from app.database import Base
+from app.db.base import Base
 
 class Region(Base):
     __tablename__ = "region"
