+++ docker-compose.production.yml
@@ -1,0 +1,7 @@
+version: '3.8'
+
+services:
+  backend:
+    environment:
+      - DATABASE_URL=postgresql://cs348:group10@db:5432/housing_db
+      - DB_ENV=productionoad
