+++ docker-compose.yml
@@ -28,7 +28,7 @@ services:
       db:
         condition: service_healthy
     environment:
-      - DATABASE_URL=postgresql://housing_user:housing_password@db:5432/housing_db
+      - DATABASE_URL=postgresql://cs348:group10@db:5432/housing_db
     volumes:
       - ./dataset:/app/dataset
     networks:
@@ -39,8 +39,8 @@ services:
     image: postgres:17
     environment:
       POSTGRES_DB: housing_db
-      POSTGRES_USER: housing_user
-      POSTGRES_PASSWORD: housing_password
+      POSTGRES_USER: cs348
+      POSTGRES_PASSWORD: group10
     ports:
       - "5432:5432"
     volumes:
@@ -48,7 +48,7 @@ services:
     networks:
       - app-network
     healthcheck:
-      test: ["CMD-SHELL", "pg_isready -U housing_user -d housing_db"]
+      test: ["CMD-SHELL", "pg_isready -U cs348 -d housing_db"]
       interval: 5s
       timeout: 5s
       retries: 5
